Refactor: 오타 수정, claeatAutomatically 속성 변경
