+++ blog/src/main/java/com/practice/blog/account/dto/request/CreateAccountRequestDto.java
@@ -1,4 +1,4 @@
-package com.practice.blog.account.dto.requset;
+package com.practice.blog.account.dto.request;
 
 import com.practice.blog.account.entity.Account;
 import jakarta.validation.constraints.NotBlank;
